Use WS.sendTextData instead of WS.sendBinaryData
We are always sending JSON data, we should thus use sendTextData.
Quoting the docs: "This should be the default choice for human-readable
text-based protocols such as JSON."

This has the added advantage of making it easier for the JS client
library, since it will get a string-type websocket frame, which can be
parsed directly with JSON.parse instead of having to decode a byte array
first.