Write to a temporary file first
And then rename the file. This is more robust than writing to the file
directly, since renaming is atomic, but writing big files is not.
If we would crash while writing the data file then the file would be
corrupted.
If we only corrupt the temporary file then we can recover gracefully.