Create an empty database if there is none yet
If we cannot find a database at the configured file path we used to
crash with a not terribly good error message.

This PR instead logs a warning and then creates an empty database
(an empty JSON object: {}). Note, that the object will only be
persisted when some changes are written to disk.