Implement file loading for Sqlite
Add a "loadSqliteFile" function that reads the "configDataFile" file
path, which must be a Sqlite file. In case the file does not exist yet,
a new database is created and the empty object ("{}") is returned as its
value.

The database is super simple right now and contains just a single row,
with a single column of type TEXT. The whole JSON value is stored in
there as a UTF-8 encoded string.