Implement writing to the Sqlite file
I added a "syncSqliteFile" function which simply issues an UPDATE.
There is only one row, so only that row will get updated.
The implementation currently still uses the TEXT type for the database,
which maps to Haskell's Text type. This is inefficient, since we then
first parse a ByteString, from disk, to Text, but then have to encode
this again in order to parse it as a JSON Value. This can be optimized
in the future by writing specialized FromJSON and ToJSON instances.