Fixed a race condition in the read_after_write.py test
We were not sending durable requests here, but we really need to do that
for this test to work correctly. Interestingly, this never went wrong
with the file backend, but did go wrong for the Sqlite backend. Probably
because the file backend is faster and thus never triggered the race
condition.