nix: Add getPkgs parameter to default.nix
This allows users to override what nixpkgs we start from. That's useful
if users want to avoid having an explosion of different versions of
Nixpkgs.

I'm not sure if this does what we want, as I suspect it might be
cumbersome to use. There might be a more idiomatic pattern for this. Is
there way to allow users to pass in a fully evaluated pkgs argument?