Terminate websocket conns when pings time out
There is no clean or built in way to do this, see
https://github.com/jaspervdj/websockets/issues/159. This implementation
works by keeping track of the last received pong, and then checking
whether the previous ping has been answered with a pong when sending a
new ping. If that isn't the case, then the connection is terminated
early through the use of `withInteruptablePingThread`.